requirements.txt
lxml>=4.9.3
click>=8.1.3
python-dotenv>=1.0.0
tqdm>=4.66.1
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-anthropic>=0.0.1
langchain-community>=0.0.10 